+++ ReaderIOS/ReaderIOS/Views/PDFViewComponents/AnnotationsComponents/ClearScreenView.swift
