+++ ReaderIOS/ReaderIOS/Views/PDFViewComponents/DigitalResourcesComponents/ContentViews.swift
