+++ ReaderIOS/ReaderIOS/Views/PDFViewComponents/TimerViews.swift
